Taken common code into an abstract class
