+++ questionGenerator/src/main/java/questionGenerator/generator/RightAnswerIsEntity.java
@@ -1,8 +1,10 @@
 package main.java.questionGenerator.generator;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
+import java.util.Random;
 
 import org.wikidata.wdtk.datamodel.implementation.ItemDocumentImpl;
 import org.wikidata.wdtk.datamodel.interfaces.Statement;
@@ -59,6 +61,26 @@ public abstract class RightAnswerIsEntity extends AbstractGenerator {
 		String[] split1 = url.split(" ");
 		String[] split2 = split1[0].split("/");
 		return split2[split2.length-1];
+	}
+
+	@Override
+	protected List<String> getWrongAnswers(String rightAnswer) {
+		Random rnd = new Random();
+		String[] entities = {"Q142", "Q183", "Q16", "Q142", "Q30", "Q408", "Q668", "Q17", "Q38", "Q159",
+		 "Q79", "Q155", "Q884", "Q414", "Q41", "Q258", "Q96", "Q843", "Q148", "Q20"};
+		List<String> result = new ArrayList<>();
+		List<Integer> used = new ArrayList<>();
+		for(int i = 0; i < 3; i++){
+				int rndnum = rnd.nextInt(entities.length);
+				String wrong = getAnswer(entities[rndnum]);
+			if(wrong.equals(rightAnswer) || used.contains(rndnum))
+				i--;
+			else{
+				result.add(wrong);
+				used.add(rndnum);
+			}
+		}
+		return result;
 	}
 
 }
