+++ questionGenerator/src/main/java/questionGenerator/generator/specificGenerators/LanguageGenerator.java
@@ -20,24 +20,24 @@ public class LanguageGenerator extends RightAnswerIsEntity {
 		return String.format(q, name);
 	}
 
-	@Override
-	protected List<String> getWrongAnswers(String rightAnswer) {
-		Random rnd = new Random();
-		String[] entities = {"Q142", "Q183", "Q16", "Q142", "Q30", "Q408", "Q668", "Q17", "Q38", "Q159",
-		 "Q79", "Q155", "Q884", "Q414", "Q41", "Q258", "Q96", "Q843", "Q148", "Q20"};
-		List<String> result = new ArrayList<>();
-		List<Integer> used = new ArrayList<>();
-		for(int i = 0; i < 3; i++){
-				int rndnum = rnd.nextInt(entities.length);
-				String wrong = getAnswer(entities[rndnum]);
-			if(wrong.equals(rightAnswer) || used.contains(rndnum))
-				i--;
-			else{
-				result.add(wrong);
-				used.add(rndnum);
-			}
-		}
-		return result;
-	}
+	// @Override
+	// protected List<String> getWrongAnswers(String rightAnswer) {
+	// 	Random rnd = new Random();
+	// 	String[] entities = {"Q142", "Q183", "Q16", "Q142", "Q30", "Q408", "Q668", "Q17", "Q38", "Q159",
+	// 	 "Q79", "Q155", "Q884", "Q414", "Q41", "Q258", "Q96", "Q843", "Q148", "Q20"};
+	// 	List<String> result = new ArrayList<>();
+	// 	List<Integer> used = new ArrayList<>();
+	// 	for(int i = 0; i < 3; i++){
+	// 			int rndnum = rnd.nextInt(entities.length);
+	// 			String wrong = getAnswer(entities[rndnum]);
+	// 		if(wrong.equals(rightAnswer) || used.contains(rndnum))
+	// 			i--;
+	// 		else{
+	// 			result.add(wrong);
+	// 			used.add(rndnum);
+	// 		}
+	// 	}
+	// 	return result;
+	// }
 
 }
